third_party/blink/web_tests/platform/mac-mac13-arm64/compositing/draws-content/webgl-simple-background-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'enable')
